[v0.9.1-dev][CI/UT][bugfix]fix v0 spec decode
